server/block: fix sign and campfire state being lost when they are saved
A sign writes its colour, glow, owner and waxed state under the keys the
Bedrock format uses, but read them back under different ones, so none of them
survived a save and load. Only the text came back, because it is the one key
both sides already agreed on. A waxed sign also became editable again, which
removes the protection waxing is used for.

A campfire wrote the cooking time of each item as a uint8 and read it as an
int16, which never matches, so the timer came back as zero and everything on
the campfire finished cooking on the first tick after the chunk was loaded.
The value did not fit in a uint8 either: a campfire cooks for 600 ticks.

+++ server/block/campfire.go
@@ -244,7 +244,7 @@ func (c Campfire) EncodeNBT() map[string]any {
 		id := strconv.Itoa(i + 1)
 		if !v.Item.Empty() {
 			m["Item"+id] = item.WriteNBT(v.Item, true)
-			m["ItemTime"+id] = uint8(v.Time.Milliseconds() / 50)
+			m["ItemTime"+id] = int16(v.Time.Milliseconds() / 50)
 		}
 	}
 	return m
